Allow a .sumi file on sumigetw command line to initiate transfer.
This paves the way for .torrent-like .sumi files that can be distributed
and linked to in order to invoke SUMI. New file format for .sumi files
(which are also used for resuming). Less code assumes sys.argv is
(transport, nick, filename).

Added a new callback message, new_xfer, moved code out of RecvReqThread().