Additional WikiText features:
Add '' emphasis for italics, also supports ''''' for bold-italics

Add "; term : definition" support, output a glossentry. Must be a space
preceding the colon (recommended by Wikipedia), and the complete entry must be
on the same line (MediaWiki supports spanning over lines, but not wt2db.)

Add : for blockquoting, nested to any depth

Add support for inline programlistings, indented with a space. This breaks the
following commands that currently allow whitespace: # * and  ----